+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -187,7 +187,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-20T11:15:45.627Z",
+    "2025-11-20T12:58:45.487Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -226,7 +226,7 @@ export const CreateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-20T11:15:45.627Z"))
+  updated: z.date().default(() => new Date("2025-11-20T12:58:45.487Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {