+++ packages/orq-rc/RELEASES.md
@@ -5219,3 +5219,13 @@ Based on:
 - [typescript v4.0.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.56 - packages/orq-rc
+
+## 2025-11-20 13:00:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.57] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.57 - packages/orq-rc